pipelines/src/pipelines/diffusion_models/separate/base_separation_dataset.py
from torch.utils.data import Dataset
import abc
from typing import Tuple
import torch
from abc import ABC

class SeparationDataset(Dataset, ABC):
    @abc.abstractmethod
    def __getitem__(self, item) -> Tuple[torch.Tensor, ...]:
        ...

    @abc.abstractmethod
    def __len__(self) -> int:
        ...

    @property
    @abc.abstractmethod
    def sample_rate(self) -> int:
        ...